adding colors to status cell
